+++ LocalPackages/RTSComponentKit/Sources/RTSComponentKit/Store/Dimensions.swift
@@ -1,0 +1,14 @@
+//
+//  File.swift
+//
+
+import Foundation
+
+public struct Dimensions {
+    public let width: Float?
+    public let height: Float?
+    public init(width: Float?, height: Float?) {
+        self.width = width
+        self.height = height
+    }
+}
